+++ src/com/intellij/idea/plugin/hybris/impex/completion/ImpexImplementationClassCompletionContributor.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -25,7 +25,6 @@ interface ImpexImplementationClassCompletionContributor {
     fun getImplementationsForClass(qualifiedName: String): Set<LookupElement>
 
     companion object {
-        fun getInstance(project: Project): ImpexImplementationClassCompletionContributor? =
-            project.getService(ImpexImplementationClassCompletionContributor::class.java)
+        fun getInstance(project: Project): ImpexImplementationClassCompletionContributor? = project.getService(ImpexImplementationClassCompletionContributor::class.java)
     }
 }