+++ src/com/intellij/idea/plugin/hybris/impex/constants/modifier/TypeModifier.kt
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -39,7 +40,7 @@ enum class TypeModifier(
     IMPEX_LEGACY_MODE("impex.legacy.mode", HybrisConstants.IMPEX_MODIFIER_BOOLEAN_VALUES),
     PROCESSOR("processor") {
         override fun getLookupElements(project: Project) = ImpexImplementationClassCompletionContributor.getInstance(project)
-            ?.getImplementationsForClass(HybrisConstants.CLASS_IMPEX_PROCESSOR)
+            ?.getImplementationsForClass(HybrisConstants.CLASS_FQN_IMPEX_PROCESSOR)
             ?: emptySet()
     };