+++ src/com/intellij/idea/plugin/hybris/lang/injection/impl/AbstractLanguageInjectorProvider.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -16,9 +16,8 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.intellij.idea.plugin.hybris.flexibleSearch.injection
+package com.intellij.idea.plugin.hybris.lang.injection.impl
 
-import com.intellij.idea.plugin.hybris.flexibleSearch.FlexibleSearchLanguage
 import com.intellij.lang.Language
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.progress.ProcessCanceledException
@@ -27,7 +26,7 @@ import com.intellij.psi.InjectedLanguagePlaces
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiLanguageInjectionHost
 
-abstract class FlexibleSearchInjectorProvider {
+abstract class AbstractLanguageInjectorProvider(private val injectLanguage: Language) {
 
     protected abstract val language: Language
 
@@ -54,7 +53,7 @@ abstract class FlexibleSearchInjectorProvider {
     ) {
         try {
             injectionPlacesRegistrar.addPlace(
-                FlexibleSearchLanguage.INSTANCE,
+                injectLanguage,
                 textRange,
                 prefix,
                 suffix
@@ -67,6 +66,6 @@ abstract class FlexibleSearchInjectorProvider {
     }
 
     companion object {
-        private val LOG = Logger.getInstance(FlexibleSearchInjectorProvider::class.java)
+        private val LOG = Logger.getInstance(AbstractLanguageInjectorProvider::class.java)
     }
 }